+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `EveOnline::ESI::CorporationOrders`
 * Remove `EveOnline::XML::CorporationMarketOrders`. Use `EveOnline::ESI::CorporationOrders` instead.
 * Remove `EveOnline::MarketOrder`
+* Add `EveOnline::ESI::Models::Bookmark`
 
 **v0.14.0**